requirements.txt
evdev
asyncio
aiohttp
python-dotenv
